Update images test class
